Separate AccountStore and TypedTokenStore
